move-NavLinks-to-separate-file-and-update-imports
